// src/store/interfaces/form-data.ts
export interface Formdata {
  name?: string;
  email?: string;
  phone?: string;
  street?: string;
  city?: string;
  state?: string;
  zipcode?: string;
  petname?: string;
  weight?: string;
  breed?: string;
  age?: number;
  gender?: string;
  activity_level?: string;
  body_shape?: string;
  diet_plan?: [];
  allergies?: string[];
  preferred_foods?: string[];
}